meta_environment/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
click==8.1.8
contourpy==1.3.0
cycler==0.12.1
exceptiongroup==1.3.1
fastapi==0.128.8
fonttools==4.60.2
h11==0.16.0
idna==3.11
importlib_resources==6.5.2
kiwisolver==1.4.7
markdown-it-py==3.0.0
matplotlib==3.9.4
mdurl==0.1.2
numpy==2.0.2
packaging==26.0
pandas==2.3.3
pillow==11.3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.20.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
pytz==2026.1.post1
rich==14.3.3
shellingham==1.5.4
six==1.17.0
starlette==0.49.3
typer==0.23.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uvicorn==0.39.0
zipp==3.23.0
openenv-core[core]>=0.2.2
requests>=2.31.0
